+++ addons/mail/static/src/main.js
@@ -19,6 +19,7 @@ import { suggestionService } from "./new/suggestion/suggestion_service";
 import { stateService } from "./new/core/state_service";
 import { chatWindowService } from "./new/chat/chat_window_service";
 import { threadService } from "./new/thread/thread_service";
+import { messageService } from "./new/thread/message_service";
 
 const messagingValuesService = {
     start() {
@@ -30,6 +31,7 @@ const serviceRegistry = registry.category("services");
 serviceRegistry.add("mail.state", stateService);
 serviceRegistry.add("mail.chat_window", chatWindowService);
 serviceRegistry.add("mail.thread", threadService);
+serviceRegistry.add("mail.message", messageService);
 serviceRegistry.add("mail.messaging", newMessagingService);
 serviceRegistry.add("mail.suggestion", suggestionService);
 serviceRegistry.add("mail.rtc", rtcService);
