+++ addons/mail/static/src/new/core/messaging.js
@@ -3,13 +3,7 @@
 import { markup, reactive } from "@odoo/owl";
 import { Deferred } from "@web/core/utils/concurrency";
 import { memoize } from "@web/core/utils/functions";
-import { registry } from "@web/core/registry";
-import {
-    prettifyMessageContent,
-    convertBrToLineBreak,
-    cleanTerm,
-    htmlToTextContentInline,
-} from "@mail/new/utils/format";
+import { cleanTerm, htmlToTextContentInline } from "@mail/new/utils/format";
 import { removeFromArray } from "@mail/new/utils/arrays";
 import { ChatWindow } from "./chat_window_model";
 import { Thread } from "./thread_model";
@@ -27,8 +21,6 @@ import { createLocalId } from "./thread_model.create_local_id";
 
 const PREVIEW_MSG_MAX_SIZE = 350; // optimal for native English speakers
 export const OTHER_LONG_TYPING = 60000;
-
-const commandRegistry = registry.category("mail.channel_commands");
 
 export const asyncMethods = [
     "fetchPreviews",
@@ -71,7 +63,8 @@ export class Messaging {
         soundEffects,
         userSettings,
         chatWindow,
-        thread
+        thread,
+        message
     ) {
         this.env = env;
         this.rpc = rpc;
@@ -81,6 +74,7 @@ export class Messaging {
         this.userSettings = userSettings;
         this.chatWindow = chatWindow;
         this.thread = thread;
+        this.message = message;
         this.nextId = 1;
         this.router = router;
         this.bus = bus;
@@ -288,32 +282,6 @@ export class Messaging {
         );
     }
 
-    /**
-     * Create a transient message, i.e. a message which does not come
-     * from a member of the channel. Usually a log message, such as one
-     * generated from a command with ('/').
-     *
-     * @param {Object} data
-     */
-    createTransientMessage(data) {
-        const { body, res_id: threadId } = data;
-        const lastMessageId = Object.values(this.state.messages).reduce(
-            (lastMessageId, message) => Math.max(lastMessageId, message.id),
-            0
-        );
-        Message.insert(
-            this.state,
-            {
-                author: this.state.partnerRoot,
-                body,
-                id: lastMessageId + 0.01,
-                is_note: true,
-                is_transient: true,
-            },
-            this.state.threads[createLocalId("mail.channel", threadId)]
-        );
-    }
-
     // -------------------------------------------------------------------------
     // process notifications received by the bus
     // -------------------------------------------------------------------------
@@ -428,18 +396,6 @@ export class Messaging {
         return Boolean(browser.Notification && browser.Notification.permission === "granted");
     }
 
-    createNotificationMessage(message, channel) {
-        const data = Object.assign(message, { body: markup(message.body) });
-        Message.insert(this.state, data, channel);
-        if (
-            !this.presence.isOdooFocused() &&
-            channel.type === "chat" &&
-            channel.chatPartnerId !== this.state.partnerRoot.id
-        ) {
-            this.notifyOutOfFocusMessage(message, channel);
-        }
-    }
-
     handleNotification(notifications) {
         console.log("notifications received", notifications);
         for (const notif of notifications) {
@@ -458,7 +414,15 @@ export class Messaging {
                         const channel = this.state.threads[createLocalId("mail.channel", id)];
                         Promise.resolve(channel ?? this.thread.joinChat(message.author.id)).then(
                             (channel) => {
-                                this.createNotificationMessage(message, channel);
+                                const data = Object.assign(message, { body: markup(message.body) });
+                                Message.insert(this.state, data, channel);
+                                if (
+                                    !this.presence.isOdooFocused() &&
+                                    channel.type === "chat" &&
+                                    channel.chatPartnerId !== this.state.partnerRoot.id
+                                ) {
+                                    this.notifyOutOfFocusMessage(message, channel);
+                                }
                                 ChatWindow.insert(this.state, {
                                     thread: channel,
                                 });
@@ -549,7 +513,7 @@ export class Messaging {
                     });
                     break;
                 case "mail.channel/transient_message":
-                    return this.createTransientMessage(
+                    return this.message.createTransient(
                         Object.assign(notif.payload, { body: markup(notif.payload.body) })
                     );
                 case "mail.link.preview/delete":
@@ -608,7 +572,7 @@ export class Messaging {
                         if (!message) {
                             continue;
                         }
-                        this.updateMessageStarredState(message, starred);
+                        this.message.updateStarred(message, starred);
                         this.state.discuss.starred.sortMessages();
                     }
                     break;
@@ -787,140 +751,6 @@ export class Messaging {
         }
     });
 
-    async postMessage(thread, body, { attachments = [], isNote = false, parentId, rawMentions }) {
-        const command = this.getCommandFromText(thread.type, body);
-        if (command) {
-            await this.thread.executeCommand(thread, command, body);
-            return;
-        }
-        let tmpMsg;
-        const subtype = isNote ? "mail.mt_note" : "mail.mt_comment";
-        const validMentions = this.getMentionsFromText(rawMentions, body);
-        const params = {
-            post_data: {
-                body: await prettifyMessageContent(body, validMentions),
-                attachment_ids: attachments.map(({ id }) => id),
-                message_type: "comment",
-                partner_ids: validMentions.partners.map((partner) => partner.id),
-                subtype_xmlid: subtype,
-            },
-            thread_id: thread.id,
-            thread_model: thread.model,
-        };
-        if (parentId) {
-            params.post_data.parent_id = parentId;
-        }
-        if (thread.type === "chatter") {
-            params.thread_id = thread.id;
-            params.thread_model = thread.model;
-        } else {
-            const tmpId = `pending${this.nextId++}`;
-            const tmpData = {
-                id: tmpId,
-                author: { id: this.state.user.partnerId },
-                attachments: attachments,
-                res_id: thread.id,
-                model: "mail.channel",
-            };
-            if (parentId) {
-                tmpData.parentMessage = this.state.messages[parentId];
-            }
-            tmpMsg = Message.insert(
-                this.state,
-                {
-                    ...tmpData,
-                    body: markup(await prettifyMessageContent(body, validMentions)),
-                },
-                thread
-            );
-        }
-        const data = await this.rpc("/mail/message/post", params);
-        const message = Message.insert(
-            this.state,
-            Object.assign(data, { body: markup(data.body) }),
-            thread
-        );
-        if (!message.isEmpty) {
-            this.rpc("/mail/link_preview", { message_id: data.id }, { silent: true });
-        }
-        if (thread.type !== "chatter") {
-            removeFromArray(thread.messages, tmpMsg.id);
-            delete this.state.messages[tmpMsg.id];
-        }
-        return message;
-    }
-
-    getMentionsFromText(rawMentions, body) {
-        const validMentions = {};
-        const partners = [];
-        const threads = [];
-        const rawMentionedPartnerIds = rawMentions.partnerIds || [];
-        const rawMentionedThreadIds = rawMentions.threadIds || [];
-        for (const partnerId of rawMentionedPartnerIds) {
-            const partner = this.state.partners[partnerId];
-            const index = body.indexOf(`@${partner.name}`);
-            if (index === -1) {
-                continue;
-            }
-            partners.push(partner);
-        }
-        for (const threadId of rawMentionedThreadIds) {
-            const thread = this.state.threads[createLocalId("mail.channel", threadId)];
-            const index = body.indexOf(`#${thread.displayName}`);
-            if (index === -1) {
-                continue;
-            }
-            threads.push(thread);
-        }
-        validMentions.partners = partners;
-        validMentions.threads = threads;
-        return validMentions;
-    }
-
-    getCommandFromText(threadType, content) {
-        if (content.startsWith("/")) {
-            const firstWord = content.substring(1).split(/\s/)[0];
-            const command = commandRegistry.get(firstWord, false);
-            if (command) {
-                const types = command.channel_types || ["channel", "chat", "group"];
-                return types.includes(threadType) ? command : false;
-            }
-        }
-    }
-
-    async updateMessage(message, body, attachments = [], rawMentions) {
-        if (convertBrToLineBreak(message.body) === body && attachments.length === 0) {
-            return;
-        }
-        const validMentions = this.getMentionsFromText(rawMentions, body);
-        const data = await this.rpc("/mail/message/update_content", {
-            attachment_ids: attachments
-                .map(({ id }) => id)
-                .concat(message.attachments.map(({ id }) => id)),
-            body: await prettifyMessageContent(body, validMentions),
-            message_id: message.id,
-        });
-        message.body = markup(data.body);
-        message.attachments.push(...attachments);
-    }
-
-    async addReaction(message, content) {
-        const messageData = await this.rpc("/mail/message/add_reaction", {
-            content,
-            message_id: message.id,
-        });
-        Message.insert(this.state, messageData, message.originThread);
-    }
-
-    async removeReaction(reaction) {
-        const messageData = await this.rpc("/mail/message/remove_reaction", {
-            content: reaction.content,
-            message_id: reaction.messageId,
-        });
-        const message = this.state.messages[reaction.messageId];
-        Message.insert(this.state, messageData, message.originThread);
-    }
-
     toggleReplyTo(message) {
         if (this.state.discuss.messageToReplyTo === message) {
             this.state.discuss.messageToReplyTo = null;
@@ -966,52 +796,10 @@ export class Messaging {
         });
     }
 
-    async toggleStar(message) {
-        await this.orm.call("mail.message", "toggle_message_starred", [[message.id]]);
-    }
-
-    async setDone(message) {
-        await this.orm.call("mail.message", "set_message_done", [[message.id]]);
-    }
-
-    updateMessageStarredState(message, isStarred) {
-        message.isStarred = isStarred;
-        if (isStarred) {
-            this.state.discuss.starred.counter++;
-            if (this.state.discuss.starred.messages.length > 0) {
-                this.state.discuss.starred.messages.push(message.id);
-            }
-        } else {
-            this.state.discuss.starred.counter--;
-            removeFromArray(this.state.discuss.starred.messages, message.id);
-        }
-    }
-
-    async deleteMessage(message) {
-        if (message.isStarred) {
-            this.state.discuss.starred.counter--;
-            removeFromArray(this.state.discuss.starred.messages, message.id);
-        }
-        message.body = "";
-        message.attachments = [];
-        return this.rpc("/mail/message/update_content", {
-            attachment_ids: [],
-            body: "",
-            message_id: message.id,
-        });
-    }
-
     async unlinkAttachment(attachment) {
         return this.rpc("/mail/attachment/delete", {
             attachment_id: attachment.id,
         });
-    }
-
-    async unstarAll() {
-        // apply the change immediately for faster feedback
-        this.state.discuss.starred.counter = 0;
-        this.state.discuss.starred.messages = [];
-        await this.orm.call("mail.message", "unstar_all");
     }
 
     /**
