+++ addons/mail/static/src/new/thread/message.xml
@@ -97,7 +97,7 @@
                             <t t-if="!env.inChatWindow or !state.isActionListSquashed">
                                 <i t-if="canToggleStar" class="o-mail-message-action-toggle-star cursor-pointer px-1 py-2 fa fa-lg"
                                     t-att-class="message.isStarred ? 'fa-star' : 'fa-star-o'"
-                                    t-on-click="() => messaging.toggleStar(props.message)"
+                                    t-on-click="() => messageService.toggleStar(props.message)"
                                     role="button" tabindex="0" title="Mark as Todo" aria-label="Mark as Todo"/>
                                 <i t-if="canBeDeleted" class="cursor-pointer px-1 py-2 fa fa-lg fa-trash" t-on-click="onClickDelete"
                                     role="button" tabindex="0" title="Delete" aria-label="Delete"/>
@@ -105,7 +105,7 @@
                                     role="button" tabindex="1" title="Edit" aria-label="Edit"/>
                                 <i t-if="canReplyTo" class="o-mail-message-action-reply-to cursor-pointer px-1 py-2 fa fa-lg fa-reply" t-on-click.stop="onClickReplyTo"
                                     role="button" tabindex="1" title="Reply" aria-label="Reply"/>
-                                <i t-if="isInInbox" class="o-mail-message-action-mark-read cursor-pointer px-1 py-2 fa fa-lg fa-check" t-on-click="() => messaging.setDone(props.message)"
+                                <i t-if="isInInbox" class="o-mail-message-action-mark-read cursor-pointer px-1 py-2 fa fa-lg fa-check" t-on-click="() => messageService.setDone(props.message)"
                                     role="button" tabindex="0" title="Mark as Read" aria-label="Mark as Read"/>
                             </t>
                         </div>
