+++ addons/mail/static/tests/helpers/webclient_setup.js
@@ -31,6 +31,7 @@ import { suggestionService } from "@mail/new/suggestion/suggestion_service";
 import { stateService } from "@mail/new/core/state_service";
 import { chatWindowService } from "@mail/new/chat/chat_window_service";
 import { threadService } from "@mail/new/thread/thread_service";
+import { messageService } from "@mail/new/thread/message_service";
 
 const ROUTES_TO_IGNORE = [
     "/web/webclient/load_menus",
@@ -102,6 +103,7 @@ function setupMessagingServiceRegistries({ loadingBaseDelayDuration = 0, messagi
         "mail.suggestion": suggestionService,
         "mail.state": stateService,
         "mail.thread": threadService,
+        "mail.message": messageService,
         "mail.chat_window": chatWindowService,
         "mail.messaging": newMessagingService,
         "mail.rtc": rtcService,
